// packages/localization/src/core/types.ts
export type Locale = {
  /**
   * An [IETF BCP 47 language tag](https://en.wikipedia.org/wiki/IETF_language_tag) with a region code.
   * @example
   * `'en-US'`, `'es-419'`, `'pl-PL'`.
   */
  languageTag: string;
  /**
   * An [IETF BCP 47 language tag](https://en.wikipedia.org/wiki/IETF_language_tag) without the region code.
   * @example
   * `'en'`, `'es'`, `'pl'`.
   */
  languageCode: string | null;
  /**
   * An [ISO 15924](https://en.wikipedia.org/wiki/ISO_15924) 4-letter script code. On Android and Web, it may be `null` if none is defined.
   * @example
   * `'Latn'`, `'Hans'`, `'Hebr'`.
   */
  languageScriptCode: string | null;
  /**
   * The region code for your device that comes from the Region setting under Language & Region on iOS, Region settings on Android and is parsed from locale on Web (can be `null` on Web).
   * @example
   * `'US'`.
   */
  regionCode: string | null;
  /**
   * The region code for the preferred language. When the language is not region-specific, it returns the same value as `regionCode`. When the language is region-specific, it returns the region code for the language (`en-CA` -> `CA`).
   * Prefer using `regionCode` for any internalization purposes.
   * @example
   * `'US'`.
   */
  languageRegionCode: string | null;
  /**
   * Currency code for the locale.
   * On iOS, it's the currency code from the `Region` setting under Language & Region, not for the current locale.
   * On Android, it's the currency specifc to the locale in the list, as there are no separate settings for selecting a region.
   * Is `null` on Web, use a table lookup based on region instead.
   * @example
   * `'USD'`, `'EUR'`, `'PLN'`.
   */
  currencyCode: string | null;
  /**
   * Currency symbol for the currency specified by `currencyCode`.
   * @example
   * `'$'`, `'€'`, `'zł'`.
   */
  currencySymbol: string | null;
  /**
   * Currency code for the locale.
   * On iOS, it's the currency code for the current locale in the list, not the device region.
   * On Android, it's equal to `currencyCode`.
   * Is `null` on Web.
   * Prefer using `currencyCode` for any internalization purposes.
   * @example
   * `'USD'`, `'EUR'`, `'PLN'`.
   */
  languageCurrencyCode: string | null;
  /**
   * Currency symbol for the currency specified by `languageCurrencyCode`.
   * Prefer using `currencySymbol` for any internalization purposes.
   * @example
   * `'$'`, `'€'`, `'zł'`.
   */
  languageCurrencySymbol: string | null;
  /**
   * Decimal separator used for formatting numbers with fractional parts.
   * @example
   * `'.'`, `','`.
   */
  decimalSeparator: string | null;
  /**
   * Digit grouping separator used for formatting large numbers.
   * @example
   * `'.'`, `','`.
   */
  digitGroupingSeparator: string | null;
  /**
   * Text direction for the locale. One of: `'ltr'`, `'rtl'`.
   */
  textDirection: 'ltr' | 'rtl';
  /**
   * The measurement system used in the locale.
   * Is `null` on Web, as user chosen measurement system is not exposed on the web and using locale to determine measurement systems is unreliable.
   * Ask for user preferences if possible.
   */
  measurementSystem: `metric` | `us` | `uk` | null;
  /**
   * The temperature unit used in the locale.
   * Returns `null` if the region code is unknown.
   */
  temperatureUnit: 'celsius' | 'fahrenheit' | null;
};